Map: code cleanup and pan to layer bounds center when layer is selected (#1975)
* cleanup CoordinateAttributeDataLayer component

* Map: pan to layer bounds center; code cleanup;
