+++ webapp/components/Map/Map.scss
@@ -26,7 +26,7 @@
     }
 
     .leaflet-popup-pane {
-      z-index: 210;
+      z-index: 600;
     }
 
     .leaflet-top.leaflet-right,
